adapt the first kinematic game to the new css and re-arrange files and script for better multiple game organisation
