+++ page/kinematic_game1.html
@@ -1,4 +1,4 @@
-<!-- Coding by Nolan-->
+<!-- Coding by Daniel-->
 <!DOCTYPE html>
 <html lang="fr">
 <head>
@@ -8,8 +8,8 @@
     <title>CCT Training - Cinématique</title>
 
     <!--======= CSS =======-->
-    <link rel="stylesheet" href="./css/main_style.css">
-    <link rel="stylesheet" href="./css/kinematic_style.css">
+    <link rel="stylesheet" href="../css/main_style.css">
+    <link rel="stylesheet" href="../css/kinematic_style.css">
 
     <!-- ====== Boxicons CSS ====== -->
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
@@ -20,7 +20,7 @@
         <header>
             <div class="image-text">
                 <span class="image">
-                    <img src="./assets/logo.png" alt="logo"> 
+                    <img src="../assets/logo.png" alt="logo"> 
                 </span>
 
                 <div class="text header-text">
@@ -43,7 +43,7 @@
                 <ul class="menu-links">
 
                     <li class="nav-link">
-                        <a href="#" id="dashboardBtn">
+                        <a href="../index.html" id="dashboardBtn">
                             <i class='bx bx-home-alt icon' ></i>
                             <span class="text nav-text">Dashboard</span>
                         </a>
@@ -57,7 +57,7 @@
                     </li>
 
                     <li class="nav-link">
-                        <a href="#" id="trainingBtn">
+                        <a href="../training.html" id="trainingBtn">
                             <i class='bx bxs-graduation icon'></i>                            
                             <span class="text nav-text">Training</span>
                         </a>
@@ -106,33 +106,23 @@
     </nav>
 
     <section class="home">
-        <div class="text">Entrainement cinematique</div>
+        <div class="text">Liaison Cinématique : Trouvez le bon mot</div>
 
-        <div class="game">
+        <main>
 
-            <img id="gameImg" src="./assets/placeholder.png" alt="placeholder">
+            <img id="gameImg" src="../assets/placeholder.png" alt="placeholder">
 
-            <div class="gameButton">
+            <div class="grid-2x2">
+                <li><button type="button" id="gameBtn1">Button 1</button></li>
 
-                <li id="">
-                    <button type="button" id="gameBtn1">Button 1</button>
-                </li>
+                <li><button type="button" id="gameBtn2">Button 2</button></li>
 
-                <li id="">
-                    <button type="button" id="gameBtn2">Button 2</button>
-                </li>
+                <li><button type="button" id="gameBtn3">Button 3</button></li>
 
-                <li id="">
-                    <button type="button" id="gameBtn3">Button 3</button>
-                </li>
-
-                <li id="">
-                    <button type="button" id="gameBtn4">Button 4</button>
-                </li>
-                
+                <li><button type="button" id="gameBtn4">Button 4</button></li>
             </div>
 
-        </div>
+        </main>
 
         <div class="modal">
             <div class="modal-content">
@@ -143,8 +133,8 @@
 
     </section>
 
-    <script src="./js/script.js"></script>
-    <script src="./js/kinematic_script.js"></script>
+    <script src="../js/script.js"></script>
+    <script type="module" src="../js/kinematic_game1.js"></script>
 
 </body>
 </html>
